+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.8.1] - 2018-09-27
 ### Added
 - CallTimeout option to set a per-statement OCI_ATTR_CALL_TIMEOUT.
 - Allow login with " AS SYSASM", as requested in #100.